core/requirements.txt
# Core dependencies for Static.news broadcast system
# AI and API
openai==1.35.3
anthropic==0.30.0
aiohttp==3.9.5
requests==2.31.0

# Audio processing
pydub==0.25.1
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1

# News and content
feedparser==6.0.11
beautifulsoup4==4.12.3
newspaper3k==0.2.8
pytz==2024.1
yfinance==0.2.37

# Text processing
nltk==3.8.1
textblob==0.18.0

# Async and scheduling
asyncio==3.4.3
aiocron==1.8
apscheduler==3.10.4

# Data handling
pydantic==2.7.4
python-dotenv==1.0.1
redis==5.0.4

# Web framework (for health checks)
fastapi==0.111.0
uvicorn==0.30.1

# Monitoring and logging
structlog==24.2.0
prometheus-client==0.20.0

# Business logic
stripe==9.8.0
sendgrid==6.11.0

# Utils
python-dateutil==2.9.0
tenacity==8.3.0
cachetools==5.3.3